chore: move duplicated functions into different files, so they can be require() and used in multiple places
