+++ pages/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Epochtal</title>
     <link rel="icon" href="/icons/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="common.css">
     <link rel="stylesheet" href="style.css">
     <script src="https://kit.fontawesome.com/fc2642e2d1.js" crossorigin="anonymous"></script>
   </head>
@@ -193,6 +194,7 @@
 
     </div>
 
+    <script src="shared.js"></script>
     <script src="tools.js"></script>
     <script src="main.js"></script>
