+++ pages/live/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/common.css">
     <link rel="stylesheet" href="/style.css">
     <link rel="stylesheet" href="/live/style.css">
     <script src="https://kit.fontawesome.com/fc2642e2d1.js" crossorigin="anonymous"></script>
@@ -59,6 +60,7 @@
 
     </div>
 
+    <script src="/shared.js"></script>
     <script src="/tools.js"></script>
     <script src="/live/main.js"></script>
