+++ pages/live/lobby/index.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Epochtal Live Lobby</title>
     <link rel="icon" href="/icons/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="/common.css">
     <link rel="stylesheet" href="/style.css">
     <link rel="stylesheet" href="/live/style.css">
     <script src="https://kit.fontawesome.com/fc2642e2d1.js" crossorigin="anonymous"></script>
@@ -115,6 +116,7 @@
 
     </div>
 
+    <script src="/shared.js"></script>
     <script src="/tools.js"></script>
     <script src="/live/lobby.js"></script>
