+++ pages/main.js
@@ -97,19 +97,7 @@ var homepageInit = async function () {
   const archives = await (await fetch("/api/archive/list")).json();
   const mapvotes = await (await fetch("/api/votes/get")).json();
 
-  const whoami = await (await fetch("/api/users/whoami")).json();
-
-  // Change the login button to a logout button if the user is logged in
-  if (whoami !== null) {
-
-    const loginButton = document.querySelector("#login-button");
-
-    loginButton.innerHTML = "Log out";
-    loginButton.onclick = function () {
-      window.location.href = '/api/auth/logout';
-    };
-
-  }
+  const whoami = await handleWhoami()
 
   // Update week number
   document.querySelector("#intro-week").innerHTML = config.number;
@@ -971,33 +959,7 @@ var homepageInit = async function () {
     showPopup("Not found", "The page you were trying to access does not exist. You've been brought back to the homepage.", POPUP_ERROR);
   }
 
-  let cliKeysControl = false;
-  let cliKeysTilde = false;
-
-  // Handle CLI popup
-  const keyDownFunc = function (e) {
-    if (e.key === "Control") cliKeysControl = true;
-    if (e.key === "`") cliKeysTilde = true;
-    if (cliKeysControl && cliKeysTilde) {
-
-      const features = "popup=yes,width=640,height=400,left=20,top=20";
-
-      const popupWindow = window.open("/admin/cli/index.html", "_blank", features);
-      if (popupWindow) popupWindow.focus();
-
-      cliKeysControl = false;
-      cliKeysTilde = false;
-
-    }
-  };
-
-  const keyUpFunc = function (e) {
-    if (e.key === "Control") cliKeysControl = false;
-    if (e.key === "`") cliKeysTilde = false;
-  };
-
-  window.addEventListener("keydown", keyDownFunc);
-  window.addEventListener("keyup", keyUpFunc);
+  handleCliPopup()
 
 };
