+++ pages/profile/index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../style.css">
+    <link rel="stylesheet" href="/common.css">
+    <link rel="stylesheet" href="/style.css">
     <link rel="stylesheet" href="style.css">
     <script src="https://kit.fontawesome.com/fc2642e2d1.js" crossorigin="anonymous"></script>
     <title>Epochtal User Profile</title>
@@ -16,7 +17,7 @@
       <div id="bg-anim-bg"></div>
       <canvas id="bg-anim"></canvas>
       <div id="bg-anim-blur"></div>
-      <script src="../bg.js"></script>
+      <script src="/bg.js"></script>
     </div>
 
     <div id="page-content">
@@ -63,7 +64,8 @@
 
     </div>
 
-    <script src="../tools.js"></script>
+    <script src="/shared.js"></script>
+    <script src="/tools.js"></script>
     <script src="main.js"></script>
 
   </body>
