+++ pages/stream/controller/index.html
@@ -5,6 +5,7 @@
     <meta charset="utf-8">
     <title>Stream UI Controller</title>
     <link rel="icon" href="/icons/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="/common.css">
     <link rel="stylesheet" href="/style.css">
     <style>
       /* Disable the "Play Selected Run" button by default */
@@ -57,6 +58,8 @@
 
     <footer class="marginy"></footer>
 
+    <script src="/stream/common.js"></script>
+    <script src="/shared.js"></script>
     <script src="/tools.js"></script>
     <script src="/stream/controller.js"></script>
