+++ pages/stream/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <title>Epochtal Tournament System</title>
     <link rel="icon" href="/icons/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="/common.css">
     <link rel="stylesheet" href="/stream/stream.css">
     <script src="https://w.soundcloud.com/player/api.js"></script>
     <script src="https://kit.fontawesome.com/fc2642e2d1.js" crossorigin="anonymous"></script>
@@ -107,6 +108,8 @@
 
     </div>
 
+    <script src="/shared.js"></script>
+    <script src="/stream/common.js"></script>
     <script src="/tools.js"></script>
     <script src="/stream/music.js"></script>
     <script src="/stream/animtext.js"></script>
