+++ pages/style.css
@@ -18,20 +18,6 @@ body {
   overflow-y: auto;
 }
 
-#bg-anim-container {
-  display: block;
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 1;
-  opacity: 1;
-  pointer-events: none;
-  will-change: opacity;
-  transition: opacity 0.5s;
-}
-
 #bg-anim-container * {
   position: absolute;
   width: 100%;
